feat: adjust validations and unique constraint errors

+++ backend/src/companies/companies.service.ts
@@ -4,15 +4,25 @@ import { CreateCompanyDto } from './dto/create-company.dto';
 import { UpdateCompanyDto } from './dto/update-company.dto';
 import { PaginatedResponseDto } from 'src/common/pagination/pagination-response.dto';
 import { Company } from '@prisma/client';
+import { cnpj } from 'cpf-cnpj-validator';
+import { ConflictException } from '@nestjs/common';
 
 @Injectable()
 export class CompaniesService {
   constructor(private readonly companiesRepo: CompaniesRepository) {}
 
   async create(dto: CreateCompanyDto) {
-    return this.companiesRepo.create(dto);
+    try {
+      return await this.companiesRepo.create(dto);
+    } catch (error: any) {
+      if ('code' in error && error.code === 'P2002') {
+        throw new ConflictException(
+          `Já existe uma empresa com o CNPJ ${cnpj.format(dto.cnpj)}`,
+        );
+      }
+      throw error;
+    }
   }
-
   async findAll(
     page: number = 1,
     page_size: number = 10,
@@ -27,10 +37,24 @@ export class CompaniesService {
   }
 
   async update(id: number, dto: UpdateCompanyDto) {
-    return this.companiesRepo.update(id, dto);
+    const company = await this.companiesRepo.findById(id);
+    if (!company) throw new NotFoundException('Empresa não encontrada');
+
+    try {
+      return this.companiesRepo.update(id, dto);
+    } catch (error: any) {
+      if ('code' in error && error.code === 'P2002') {
+        throw new ConflictException(
+          `Já existe uma empresa com o CNPJ ${cnpj.format(dto.cnpj!)}`,
+        );
+      }
+      throw error;
+    }
   }
 
   async delete(id: number) {
+    const company = await this.companiesRepo.findById(id);
+    if (!company) throw new NotFoundException('Empresa não encontrada');
     return this.companiesRepo.delete(id);
   }
 }
